Moved pika tests from multicore to pika
