+++ tests/tests_changed.txt
@@ -43,3 +43,8 @@ multicore_test
 pika_publisher (NEED UNITTEST)
 pika_subsriber (NEED UNITTEST)
 couldAMQP_test
+
+------
+5th June
+- removed VM tests
+- moved pika tests from multicore_test.py to pike_test.py
